test(network): move sync tests into testing crate
