+++ testing/Cargo.toml
@@ -27,6 +27,7 @@ snarkos-posw = { path = "../posw", version = "1.1.4" }
 snarkos-storage = { path = "../storage", version = "1.1.4"}
 
 anyhow = { version = "1.0.33" }
+bincode = { version = "1.3.1" }
 once_cell = { version = "1.5.2" }
 parking_lot = { version = "0.11.1" }
 rand = { version = "0.7.0" }
